Fixed bulk insert close bug
In the CleanUp function there was a bug where we check for BoundParams
to make sure that it does not contain a key 'SQLConnection'.  If it
doesn't contain such a key then we close and cleanup.  The actual key
that gets passed for this to work is SQLiteConnection so made a slight
edit to change this and tested and the connection stays open after a
bulk insert where SQLiteConnection is passed as a parameter.